Changed way plots are rendered
Previously all plots were re-generated when an input (either measurand selection or date range) was changed, causing unnecessary loading times.

Now, all plots are generated when the date range is changed, and hidden/displayed by CSS when the measurand checkboxes are interacted with. This results in instantaneous addition/removal of measurand plots with minimal additional overhead