Feature/plotting met (#10)
* Added met plots

Have got temperature and relative humidity plots working, as well as a combined NOx plot. The checkbox values are now hardcoded, which should reduce some of the page load time.

* Added wind rose

Using openair package to display wind rose. This adds an additional dependency but is a well established package

* Removed openair logging messages

* Hardcoded x-axis limits

Guarantees that all time-series are on the same scale and thus in sync

* White header + colour logos

* Wind speed is a separate time series

* Reordered plots

* Feature/separate aq met (#9)

* First attempt at box layout

* Working implementation

* Met plots smaller than aq

* Missing import