Single wind rose + checkbox
It's very challenging to organise the layout of 8 wind roses. It is much easier to have a single wind rose with a checkbox to select the variable to colour it by. There is also less benefit from having all wind roses visible at the same time, as unlike with time-series, users are less likely to want to compare wind roses.